:hammer: Restructure ern-container-gen
