+++ ern-container-gen/src/generators/android/hull/gradle.properties
