+++ ern-container-gen/src/generators/android/hull/lib/src/main/AndroidManifest.xml
