+++ ern-container-gen/src/generators/android/hull/lib/src/main/assets/README.md
