+++ ern-container-gen/src/generators/android/hull/lib/src/main/java/com/facebook/react/bridge/SafeActivityStarter.java
