+++ ern-container-gen/src/generators/android/hull/lib/src/main/java/com/walmartlabs/ern/container/miniapps/MiniAppsConfig.java
