+++ ern-container-gen/src/generators/android/hull/lib/src/main/res/README.md
