+++ ern-container-gen/src/generators/android/hull/settings.gradle
