+++ ern-container-gen/src/generators/android/templates/MiniAppActivity.mustache
