+++ ern-container-gen/src/generators/ios/hull/ElectrodeContainer/NSBundle+frameworkBundle.m
