+++ ern-container-gen/src/publishers/GithubPublisher.js
