+++ ern-container-gen/src/publishers/MavenPublisher.js
