+++ ern-container-gen/src/publishers/Publisher.js
